+++ components/eamxx/src/physics/aux_tracers/water_isotopes/eamxx_water_isotopes_process_interface.hpp
@@ -1,7 +1,7 @@
 #ifndef SCREAM_WATER_ISOTOPES_HPP
 #define SCREAM_WATER_ISOTOPES_HPP
 
-#include "eamxx_water_tracers_process_interface.hpp"
+#include "physics/aux_tracers/water_tracers/eamxx_water_tracers_process_interface.hpp"
 #include "ekat/ekat_parameter_list.hpp"
 
 #include <string>
